add packCmd
